+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.test.tsx","./src/app.tsx","./src/home.tsx","./src/declaration.d.ts","./src/index.tsx","./src/react-app-env.d.ts","./src/setuptests.ts","./src/components/results.tsx","./src/components/selectabletile.tsx","./src/components/sequencetile.tsx","./src/games/associationrecall.tsx","./src/games/memorisesequence.tsx","./src/games/standardquiz.tsx","./src/hooks/data-fetch/usefetchcountrycapitalpairs.tsx","./src/hooks/game-state/useassociationrecallstate.tsx","./src/hooks/game-state/usememorysequencegamestate.tsx","./src/layout/layout.tsx","./src/lib/getgrade.ts","./src/lib/randomselection.ts","./src/lib/snaketocamel.ts","./src/types/types.ts"],"version":"5.6.2"}
+{"root":["./src/app.test.tsx","./src/app.tsx","./src/home.tsx","./src/declaration.d.ts","./src/index.tsx","./src/react-app-env.d.ts","./src/setuptests.ts","./src/components/results.tsx","./src/components/selectabletile.tsx","./src/components/sequencetile.tsx","./src/games/associationrecall.tsx","./src/games/memorisesequence.tsx","./src/games/standardquiz.tsx","./src/hooks/data-fetch/usefetchcountrycapitalpairs.tsx","./src/hooks/game-state/useassociationrecallstate.tsx","./src/hooks/game-state/usememorysequencegamestate.tsx","./src/layout/layout.tsx","./src/lib/getgrade.ts","./src/lib/randomselection.ts","./src/lib/snaketocamel.ts","./src/types/constants.ts","./src/types/types.ts"],"version":"5.6.2"}
